third_party/blink/web_tests/crypto/subtle/hkdf/deriveKey-expected.txt

Test that keys can be derived from an HKDF key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS derivedKey.type is "secret"
PASS typeof derivedKey.extractable is 'boolean'
PASS derivedKey.extractable is false
PASS derivedKey.algorithm.name is "AES-GCM"
PASS derivedKey.usages.join(',') is "encrypt"

Try to derive an HKDF key...
error is: OperationError: No length was specified for the HKDF Derive Bits operation.
PASS emptyKey.type is "secret"
PASS typeof emptyKey.extractable is 'boolean'
PASS emptyKey.extractable is false
PASS emptyKey.algorithm.name is "HKDF"
PASS emptyKey.usages.join(',') is "deriveKey"
PASS derivedKey.type is "secret"
PASS typeof derivedKey.extractable is 'boolean'
PASS derivedKey.extractable is false
PASS derivedKey.algorithm.name is "AES-GCM"
PASS derivedKey.usages.join(',') is "encrypt"
PASS successfullyParsed is true

TEST COMPLETE
